+++ lib/tfl_api_client.rb
@@ -25,6 +25,8 @@
 
 require 'tfl_api_client/version'
 require 'tfl_api_client/client'
+require 'tfl_api_client/accident_stats'
+require 'tfl_api_client/air_quality'
 require 'tfl_api_client/bike_point'
 require 'tfl_api_client/cycle'
 require 'tfl_api_client/exceptions'