+++ spec/unit/cycle_spec.rb
@@ -26,7 +26,7 @@
 require_relative '../spec_helper'
 
 describe TflApi::Client::Cycle do
-  let!(:client) { TflApi::Client.new(app_id: 123, app_key: 456) }
+  let!(:client) { test_client }
   let!(:cycle) { TflApi::Client::Cycle.new(client) }
   let(:sample_response) do
     [{